backend/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-multipart==0.0.9
pymupdf==1.24.10
python-pptx==0.6.23
openai==1.12.0
aiosqlite==0.19.0
python-dotenv==1.0.1
